backend/auth-login/requirements.txt
psycopg2-binary==2.9.9
bcrypt==4.1.2
PyJWT==2.8.0
pydantic==2.5.0
email-validator==2.1.0
